pkp/pkp-lib#12509 replace store with composable
